chore: remove redundant dbtype package
This wasn't necessary and forked how we do DB types.
