+++ coderd/database/dbtype/dbtype.go
@@ -1,30 +1,1 @@
 package dbtype
-
-import (
-	"database/sql/driver"
-	"encoding/json"
-
-	"golang.org/x/xerrors"
-)
-
-type StringMap map[string]string
-
-func (m *StringMap) Scan(src interface{}) error {
-	if src == nil {
-		return nil
-	}
-	switch src := src.(type) {
-	case []byte:
-		err := json.Unmarshal(src, m)
-		if err != nil {
-			return err
-		}
-	default:
-		return xerrors.Errorf("unsupported Scan, storing driver.Value type %T into type %T", src, m)
-	}
-	return nil
-}
-
-func (m StringMap) Value() (driver.Value, error) {
-	return json.Marshal(m)
-}
